api/requirements.txt
Django~=4.2.2
psycopg2-binary~=2.9.6
djangorestframework~=3.14.0
drf_yasg~=1.21.6